requirements.txt
python-binance>=1.0.19
openai>=1.3.8
notion-client>=2.2.1
pandas>=2.2.0
requests>=2.31.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
supabase>=2.17.0 